+++ bootstrap/providers.php
@@ -6,6 +6,7 @@ use App\Auth\Infrastructure\Providers\AuthServiceProvider;
 use App\Catalog\Infrastructure\Providers\CatalogServiceProvider;
 use App\Listing\Infrastructure\Providers\ListingServiceProvider;
 use App\Location\Infrastructure\Providers\LocationServiceProvider;
+use App\Media\Infrastructure\Providers\MediaServiceProvider;
 use App\Shared\Infrastructure\Providers\AppServiceProvider;
 use App\Shared\Infrastructure\Providers\ConsoleServiceProvider;
 use App\Shared\Infrastructure\Providers\EventServiceProvider;
@@ -17,6 +18,7 @@ return [
     CatalogServiceProvider::class,
     ListingServiceProvider::class,
     LocationServiceProvider::class,
+    MediaServiceProvider::class,
     AppServiceProvider::class,
     ConsoleServiceProvider::class,
     EventServiceProvider::class,